(y60) fixed warning

+++ netsrc/SDL_ttf2/SDL_ttf.h
@@ -87,7 +87,7 @@ extern DECLSPEC double TTF_Kerning(TTF_Font * theFont, const Uint16 theFirstChar
 /* font tracking  [ART+COM Patch] */
 extern DECLSPEC void TTF_SetTracking(float theTracking);
 /* get last rendered glyph x position [ART+COM Patch] */
-extern DECLSPEC float* TTF_getCurrentGlyphXPositions();
+extern DECLSPEC int* TTF_getCurrentGlyphXPositions();
 extern DECLSPEC int TTF_getCurrentGlyphXPositionsCount();
 
 /* Open a font file and create a font of the specified point size.
